Win32 type fixes and setup update.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@885 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/endpoints/mod_iax/mod_iax.c
@@ -620,7 +620,7 @@ static switch_status channel_read_frame(switch_core_session *session, switch_fra
 
 		if (timeout > -1) {
 			elapsed = (unsigned int)((switch_time_now() - started) / 1000);
-			if (elapsed >= timeout) {
+			if (elapsed >= (unsigned int)timeout) {
 				return SWITCH_STATUS_SUCCESS;
 			}
 		}
